Move coach initialstate and mutations to own modules
