+++ kolibri/plugins/coach/assets/src/state/initialState.js
@@ -1,0 +1,10 @@
+export default {
+  pageName: '',
+  pageState: {},
+  classId: null,
+  className: null,
+  classList: [],
+  classCoaches: [],
+  currentClassroom: {},
+  busy: false,
+};
